tasktracker/pyproject.toml
[tool.poetry]
name = "tasktracker"
version = "0.0.1"
description = "Task tracker for popugs"
authors = ["abolychev <bolychev.anton@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nats-py = "^2.7.2"
fastapi = {extras = ["all"], version = "^0.110.0"}
sqlmodel = "^0.0.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
